requirements.txt
streamlit==1.41.0
streamlit-aggrid==0.3.4
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
validators==0.28.0